requirements.txt
aiogram==2.25.2
google-generativeai
python-dotenv